nit: move getrawtransaction tests in usual integration tests
